backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.41
databases[sqlite]==0.6.2
aiosqlite==0.20.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
httpx==0.24.1
websockets==11.0.3
python-multipart==0.0.6
aiofiles==23.1.0
openai==1.63.1
google-cloud-speech==2.20.0
soundfile==0.12.1
numpy>=2.0.0
